mod: use size_t for field size with FOFS macro, refs #157
This likely would have never come up as an issue but in theory if the offsets of gentity_t would grow beyond UINT32_MAX, this would break on 64-bit platforms.

Also simplify the FOFS macro itself to just use built-in offsetof macro.

+++ src/tvgame/tvg_utils.c
@@ -225,7 +225,7 @@ void TVG_TeamCommand(team_t team, const char *cmd)
  * @param[in] match
  * @return
  */
-gentity_t *TVG_Find(gentity_t *from, int fieldofs, const char *match)
+gentity_t *TVG_Find(gentity_t *from, size_t fieldofs, const char *match)
 {
 	char      *s;
 	gentity_t *max = &g_entities[level.num_entities];
